add tests

+++ src/meltano/api/controllers/settings.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, request
 from flask_restful import Api, Resource, fields, marshal, marshal_with
 from flask_security import roles_required
 from flask_principal import Permission, Need
+from werkzeug.exceptions import Forbidden
 from sqlalchemy.orm import joinedload
 from meltano.api.security import api_auth_required, users
 from meltano.api.models.security import db, User, Role, RolesUsers, RolePermissions
@@ -133,6 +134,9 @@ class RolesResource(Resource):
             role = Role.query.filter_by(name=role["name"]).one()
 
             if not user:
+                if role == "admin":
+                    return "The `admin` role cannot be deleted.", 403
+
                 # delete the role
                 db.session.delete(role)
             else:
@@ -196,5 +200,5 @@ class RolePermissionsResource(Resource):
 
 
 settingsApi.add_resource(AclResource, "/acl")
-settingsApi.add_resource(RolePermissionsResource, "/acl/roles/permissions")
-settingsApi.add_resource(RolesResource, "/acl/roles")
+settingsApi.add_resource(RolePermissionsResource, "/acl/roles/permissions", endpoint="role_permissions")
+settingsApi.add_resource(RolesResource, "/acl/roles", endpoint="roles")
